Implemented a line-buffering system.
The tty and each session are allocated a circular buffer.  When a record delimiter
is detected, writes are attempted.  If writes to tty partially succeed, more data is read,
time passes and the pending writes are reattempted.  The tty to session writes are
copied to a scratch buffer and written to each session from there.

Signed-off-by: Russell Senior <russell@personaltelco.net>